feature (Unreal) Ore Support (#12941) eeb280d7f9 * started getting ore to build with unreal
* building with ore but not using ore

* now subclass context and stub functions

* started implementing rhi ore backend

* make work with rhi

* updated view desc to be rcp for texture

* builds in unreal

* added asset id flag

* started using new internal listener to intercept shader assetts for ore shaders

* shaders now apear to compile

* formatting

* updating to new hlslblob type

* pppipelines now potentially setup

* all ore is stbbued

* revert to master

* format

* no exceptions on load

* re enable canvas draw

* reseet back to master

* reset to master

* updated to use new ore dynamic classes

* revert

* implemented buffer upload.

* started updating to new format

* wip

* started implementing new rstb

* wip ore

* wip ore

* basics of ore seems to be working

* asset cooking started with versioned asstes

* run PIE ore nows work

* ore working in packaged build for now

* shadow maps now working. also added better logging for rhi breadcrumbs

* cleaned up logs

* ore rhi now works across d3d11-12 nad vulkan

* updated to build with master

* render targets now draw canvas

* console 1 working

* hid all console references

* builds for xbox ps5

* text

* one

* two

* now works

* proper obfuscation

* final

* platform specific files

* compiles for 5.7

* resolve shader

* metal now works

* fixed open widget in unreal editor

* comments cleanup

* debug logging

* removed dead code, fixed some comments

* re worked previous workaround condition to match new way

* missing files

* fix some compile errors

* premake changes

* more premake changes

* more premgake

* wagyu premake

* more premake

* correct primitive count

* set primitive type

* fixed pointer issue

+++ cg_renderer/src/cg_factory.cpp
@@ -609,7 +609,8 @@ AutoCF<CGImageRef> CGRenderer::FlipCGImageInY(AutoCF<CGImageRef> image)
     auto w = CGImageGetWidth(image);
     auto h = CGImageGetHeight(image);
     AutoCF space = CGColorSpaceCreateDeviceRGB();
-    auto info = kCGBitmapByteOrder32Big | kCGImageAlphaPremultipliedLast;
+    auto info = static_cast<uint>(kCGBitmapByteOrder32Big) |
+                static_cast<uint>(kCGImageAlphaPremultipliedLast);
     AutoCF ctx = CGBitmapContextCreate(nullptr, w, h, 8, 0, space, info);
     CGContextConcatCTM(ctx, CGAffineTransformMake(1, 0, 0, -1, 0, h));
     CGContextDrawImage(ctx, CGRectMake(0, 0, w, h), image);
